requirements.txt
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
SQLAlchemy>=2.0.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
streamlit>=1.28.0
python-dotenv>=1.0.0
openai>=1.0.0
anthropic>=0.7.0
matplotlib>=3.7.0
plotly>=5.17.0
joblib>=1.3.0
pytest>=7.4.0
httpx>=0.25.0

